# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Praktikum Pertemuan 12",
        "description": "Dzaki Achmad Abimanyu",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "dzakiachmad550@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/getgallery": {
            "get": {
                "tags": [
                    "Ambil Data Gallery"
                ],
                "summary": "Succes",
                "description": "Apakah Berhasil?",
                "operationId": "GetGallery",
                "responses": {
                    "default": {
                        "description": "Berhasil!"
                    }
                }
            }
        },
        "/api/postGallery": {
            "post": {
                "tags": [
                    "Up Gambar"
                ],
                "summary": "Gambar TerUpload",
                "description": "Endpoint untuk upload gambar.",
                "operationId": "postGallery",
                "requestBody": {
                    "description": "Data untuk upload gambar",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Judul Upload",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Deskripsi Gambar",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "File Gambar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}